Generalise and reuse `Manifest` and associated
