+++ upki/src/revocation/index.rs
@@ -12,10 +12,10 @@ use std::path::PathBuf;
 use clubcard_crlite::TimestampInterval;
 use clubcard_crlite::{CRLiteClubcard, CRLiteStatus, LogId, Timestamp};
 
-#[cfg(feature = "__fetch")]
-use super::Manifest;
 use super::{Error, RevocationCheckInput, RevocationStatus};
 use crate::Config;
+#[cfg(feature = "__fetch")]
+use crate::data::Manifest;
 
 /// Binary-encoded index of universe metadata for all filters in a manifest.
 ///
